requirements.txt
Pillow==9.5.0
google-genai==1.23.0
moviepy==1.0.3
numpy==2.3.1
pydantic==2.11.7
python-dotenv==1.1.1
rich==14.0.0